split object3d.py
