+++ rosys/geometry/__init__.py
@@ -1,9 +1,11 @@
 from .line import Line
 from .line_segment import LineSegment
-from .object3d import Frame3d, Point3d, Pose3d
+from .object3d import Object3d
 from .point import Point
+from .point3d import Point3d
 from .polygon import Polygon
 from .pose import Pose, PoseStep
+from .pose3d import Frame3d, Pose3d
 from .prism import Prism
 from .rectangle import Rectangle
 from .rotation import Rotation
@@ -16,6 +18,7 @@ __all__ = [
     'LineSegment',
     'Point',
     'Point3d',
+    'Object3d',
     'Polygon',
     'Pose',
     'PoseStep',
