+++ rosys/geometry/frame3d_registry.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .pose3d import Frame3d
+
+frame_registry: dict[str, Frame3d] = {}
